+++ src/main/java/mum/pmp/mstore/repository/profile/RoleRepository.java
@@ -1,4 +1,4 @@
-package mum.pmp.mstore.repository.security;
+package mum.pmp.mstore.repository.profile;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
